Extract apiSpec model
