+++ packages/insomnia-inso/src/commands/lint-specification.js
@@ -2,7 +2,7 @@
 import { Spectral } from '@stoplight/spectral';
 import type { GlobalOptions } from '../util';
 import { loadDb } from '../db';
-import { loadApiSpec, promptApiSpec } from '../db/prompts';
+import { loadApiSpec, promptApiSpec } from '../db/models/api-spec';
 
 export type LintSpecificationOptions = GlobalOptions;
